Added possibility to define server time diff for some plugins
Config string for mod_chat_history_sqlite and mod welcome can contain server_time_diff variable. This is an integer value to modify time which is sent to the user in motd or the time when message was sent.

Note: Neither UTC nor DST needed, local server time is taken as the base (already counting with these two).